chore: rector SetList CODE_QUALITY

+++ demosplan/DemosPlanCoreBundle/Logic/Statement/StatementDeleter.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 
 namespace demosplan\DemosPlanCoreBundle\Logic\Statement;
 
+use demosplan\DemosPlanCoreBundle\Entity\Statement\ConsultationToken;
 use DemosEurope\DemosplanAddon\Contracts\Events\StatementPreDeleteEventInterface;
 use DemosEurope\DemosplanAddon\Contracts\MessageBagInterface;
 use DemosEurope\DemosplanAddon\Contracts\PermissionsInterface;
@@ -166,7 +167,7 @@ class StatementDeleter
             if ($allowedToDelete) {
                 try {
                     // Prohibit deletion if a consultation token exists for this statement
-                    if (null !== $this->consultationTokenService->getTokenForStatement($statement)) {
+                    if ($this->consultationTokenService->getTokenForStatement($statement) instanceof ConsultationToken) {
                         throw new DemosException('error.delete.statement.consultation.token', 'Statement '.DemosPlanTools::varExport($statementId, true).' has an associated consultation token.');
                     }
                     if ($canTransaction) {
@@ -179,14 +180,11 @@ class StatementDeleter
                     $deleted = $this->statementRepository->delete($statementId);
                     // add report:
                     try {
-                        if (true === $deleted) {
+                        if ($deleted) {
                             $originalStatement = $statement->getOriginal();
-                            if ($originalStatement instanceof Statement) {
-                                if ($this->permissions->hasPermission('feature_auto_delete_original_statement')
-                                    && $originalStatement->getChildren()->isEmpty()
-                                ) {
-                                    $this->deleteOriginalStatement($originalStatement);
-                                }
+                            if ($originalStatement instanceof Statement && ($this->permissions->hasPermission('feature_auto_delete_original_statement')
+                                && $originalStatement->getChildren()->isEmpty())) {
+                                $this->deleteOriginalStatement($originalStatement);
                             }
                         }
                     } catch (Exception $e) {
